Create an autostart file if user systemd service cannot be created (#3)
- create ~/.config/autostart/<...>.desktop file in case user systemd service cannot be created or has been
  explicitly skipped
- separated out the env file creation which should be invoked for both systemd service file creation and is
  also used by the autostart file; also added -A/--no-autostart-file option to skip it
- remove this file if present in ybox-destroy; added a "-K/--keep-files" argument to ybox-destroy that will
  skip deleting the unshared root directory, configs/scripts dir and status file
- fail if a container name with/without "ybox-" prefix is present whose service file names will overlap with
  that of the container being created
- fix some missing handling in the unshared root case by changing the code to use a common code
  path for both the shared/unshared root cases
- check for existing service files and rename them if present

+++ src/ybox/run/destroy.py
@@ -39,7 +39,7 @@ def main_argv(argv: list[str]) -> None:
     container_name = args.container_name
 
     # check if there is a systemd service for the container
-    ybox_svc_prefix = ybox_systemd_service_prefix(container_name)
+    ybox_svc_prefix = ybox_service_prefix(container_name)
     ybox_svc = f"{ybox_svc_prefix}.service"
     systemctl = check_systemd_service_present(ybox_svc)
 
@@ -65,6 +65,9 @@ def main_argv(argv: list[str]) -> None:
         run_command([systemctl, "--user", "disable", ybox_svc], exit_on_error=False)
         Path(systemd_dir, ybox_svc).unlink(missing_ok=True)
         run_command([systemctl, "--user", "daemon-reload"], exit_on_error=False)
+    else:
+        # remove the autostart file if present
+        Path(f"{env.home}/.config/autostart/{ybox_svc_prefix}.desktop").unlink(missing_ok=True)
     # remove the .env file
     Path(env.config_dir, f"{container_name}.env").unlink(missing_ok=True)
     # also try to delete the .env file from the old location
@@ -80,35 +83,36 @@ def main_argv(argv: list[str]) -> None:
         if (runtime_conf := state.get_container_configuration(container_name)) is None:
             print_error(f"No entry found for '{container_name}' in the state database")
             sys.exit(1)
-        conf = StaticConfiguration(env, runtime_conf.distribution, container_name)
-        print_notice("Removing container configuration files and scripts")
-        shutil.rmtree(conf.configs_dir, ignore_errors=True)
-        shutil.rmtree(conf.scripts_dir, ignore_errors=True)
-        Path(conf.status_file).unlink(missing_ok=True)
-        if not runtime_conf.shared_root:
-            # remove non-shared root directory
-            unshared_root = conf.unshared_root()
-            if os.path.isdir(unshared_root):
-                print_notice(f"Removing unshared root directory {unshared_root}")
-                delete_container_directory(unshared_root, env)
-            # remove the container specific image
-            container_image = conf.box_image(False)
-            print_notice(f"Removing unshared container image {container_image}")
-            run_command([docker_cmd, "image", "rm", container_image], exit_on_error=False,
-                        error_msg="removing unshared container image")
-        # delete all container related files if required
-        if args.delete_files:
-            box_conf = get_parsed_box_conf(runtime_conf.ini_config)
-            assert box_conf is not None
-            home_dir = box_conf["base"]["home"]
-            if os.path.isdir(home_dir):
-                print_notice(f"Removing home directory {home_dir}")
-                delete_container_directory(home_dir, env)
-            container_dir = f"{env.data_dir}/{container_name}"
-            print_notice(f"Removing container directory {container_dir}")
-            delete_container_directory(container_dir, env)
-        state.unregister_container(container_name)
-        remove_orphans_from_db(valid_containers, state)
+        if not args.keep_files:
+            conf = StaticConfiguration(env, runtime_conf.distribution, container_name)
+            print_notice("Removing container configuration files and scripts")
+            shutil.rmtree(conf.configs_dir, ignore_errors=True)
+            shutil.rmtree(conf.scripts_dir, ignore_errors=True)
+            Path(conf.status_file).unlink(missing_ok=True)
+            if not runtime_conf.shared_root:
+                # remove non-shared root directory
+                unshared_root = conf.unshared_root()
+                if os.path.isdir(unshared_root):
+                    print_notice(f"Removing unshared root directory {unshared_root}")
+                    delete_container_directory(unshared_root, env)
+                # remove the container specific image
+                container_image = conf.box_image(False)
+                print_notice(f"Removing unshared container image {container_image}")
+                run_command([docker_cmd, "image", "rm", container_image], exit_on_error=False,
+                            error_msg="SKIP")
+            # delete all container related files if required
+            if args.delete_files:
+                box_conf = get_parsed_box_conf(runtime_conf.ini_config)
+                assert box_conf is not None
+                home_dir = box_conf["base"]["home"]
+                if os.path.isdir(home_dir):
+                    print_notice(f"Removing home directory {home_dir}")
+                    delete_container_directory(home_dir, env)
+                container_dir = f"{env.data_dir}/{container_name}"
+                print_notice(f"Removing container directory {container_dir}")
+                delete_container_directory(container_dir, env)
+            state.unregister_container(container_name)
+            remove_orphans_from_db(valid_containers, state)
 
 
 def parse_args(argv: list[str]) -> argparse.Namespace:
@@ -123,13 +127,17 @@ def parse_args(argv: list[str]) -> argparse.Namespace:
                         help="force destroy the container using SIGKILL if required")
     parser.add_argument("-D", "--delete-files", action="store_true",
                         help="remove all files of the container including the home directory")
+    parser.add_argument("-K", "--keep-files", action="store_true",
+                        help="keep the files of the container including the storage location of a "
+                        "non-shared root one, configs and scripts directories, non-shared image; "
+                        "systemd/autostart service and env files are still deleted")
     parser.add_argument("container_name", type=str, help="name of the active ybox")
     parser_version_check(parser, argv)
     return parser.parse_args(argv)
 
 
-def ybox_systemd_service_prefix(container_name: str) -> str:
-    """systemd service name prefix for given ybox container name"""
+def ybox_service_prefix(container_name: str) -> str:
+    """service name prefix to be used by systemd/autostart for given ybox container name"""
     return container_name if container_name.startswith("ybox-") else f"ybox-{container_name}"
